[refactor]: merge FocusRepository + FocusWriter into FocusStore
Reads (list) and writes (create/delete focus, append/delete task) of the same
markdown directory format were split across two modules with a coordinated
schema. Merge into one FocusStore trait + MarkdownFocusStore impl so the
markdown layout is owned in one place. atomic.rs stays as the sealed
crash-safe I/O utility.

Collapses two Arc<dyn ...> deps in router/serve/applier/ui_bridge/app into a
single Arc<dyn FocusStore>. Single FocusStoreError replaces RepositoryError +
WriterError. ProposalDispatcher::from_writer renamed from_store.
