+++ crates/http-api/src/applier.rs
@@ -1,26 +1,26 @@
 use std::sync::Arc;
 
 use adhd_ranch_domain::{Proposal, ProposalKind};
-use adhd_ranch_storage::{FocusWriter, WriterError};
+use adhd_ranch_storage::{FocusStore, FocusStoreError};
 
 #[derive(Debug)]
 pub enum ApplyError {
-    Writer(WriterError),
+    Store(FocusStoreError),
 }
 
 impl std::fmt::Display for ApplyError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
-            Self::Writer(e) => write!(f, "apply: {e}"),
+            Self::Store(e) => write!(f, "apply: {e}"),
         }
     }
 }
 
 impl std::error::Error for ApplyError {}
 
-impl From<WriterError> for ApplyError {
-    fn from(e: WriterError) -> Self {
-        Self::Writer(e)
+impl From<FocusStoreError> for ApplyError {
+    fn from(e: FocusStoreError) -> Self {
+        Self::Store(e)
     }
 }
 
@@ -34,12 +34,12 @@ pub trait ProposalApplier: Send + Sync {
 }
 
 pub struct AddTaskApplier {
-    writer: Arc<dyn FocusWriter>,
+    store: Arc<dyn FocusStore>,
 }
 
 impl AddTaskApplier {
-    pub fn new(writer: Arc<dyn FocusWriter>) -> Self {
-        Self { writer }
+    pub fn new(store: Arc<dyn FocusStore>) -> Self {
+        Self { store }
     }
 }
 
@@ -50,7 +50,7 @@ impl ProposalApplier for AddTaskApplier {
                 target_focus_id,
                 task_text,
             } => {
-                self.writer.append_task(target_focus_id, task_text)?;
+                self.store.append_task(target_focus_id, task_text)?;
                 Ok(AppliedOutcome {
                     target: Some(target_focus_id.clone()),
                 })
@@ -61,19 +61,19 @@ impl ProposalApplier for AddTaskApplier {
 }
 
 pub struct NewFocusApplier {
-    writer: Arc<dyn FocusWriter>,
+    store: Arc<dyn FocusStore>,
     clock: Arc<dyn Fn() -> String + Send + Sync>,
     id_gen: Arc<dyn Fn() -> String + Send + Sync>,
 }
 
 impl NewFocusApplier {
     pub fn new(
-        writer: Arc<dyn FocusWriter>,
+        store: Arc<dyn FocusStore>,
         clock: Arc<dyn Fn() -> String + Send + Sync>,
         id_gen: Arc<dyn Fn() -> String + Send + Sync>,
     ) -> Self {
         Self {
-            writer,
+            store,
             clock,
             id_gen,
         }
@@ -86,7 +86,7 @@ impl ProposalApplier for NewFocusApplier {
             ProposalKind::NewFocus { new_focus } => {
                 let id = (self.id_gen)();
                 let created_at = (self.clock)();
-                let slug = self.writer.create_focus(new_focus, &id, &created_at)?;
+                let slug = self.store.create_focus(new_focus, &id, &created_at)?;
                 Ok(AppliedOutcome { target: Some(slug) })
             }
             _ => unreachable!("NewFocusApplier called with non-new_focus kind"),
@@ -121,14 +121,14 @@ impl ProposalDispatcher {
         }
     }
 
-    pub fn from_writer(
-        writer: Arc<dyn FocusWriter>,
+    pub fn from_store(
+        store: Arc<dyn FocusStore>,
         clock: Arc<dyn Fn() -> String + Send + Sync>,
         id_gen: Arc<dyn Fn() -> String + Send + Sync>,
     ) -> Self {
         Self::new(
-            Arc::new(AddTaskApplier::new(writer.clone())),
-            Arc::new(NewFocusApplier::new(writer, clock, id_gen)),
+            Arc::new(AddTaskApplier::new(store.clone())),
+            Arc::new(NewFocusApplier::new(store, clock, id_gen)),
             Arc::new(DiscardApplier),
         )
     }
@@ -151,7 +151,7 @@ impl ProposalDispatcher {
 mod tests {
     use super::*;
     use adhd_ranch_domain::{NewFocus, ProposalId, ProposalKind};
-    use adhd_ranch_storage::MarkdownFocusWriter;
+    use adhd_ranch_storage::MarkdownFocusStore;
     use std::fs;
     use tempfile::TempDir;
 
@@ -166,9 +166,9 @@ mod tests {
     }
 
     fn dispatcher(focuses_root: &std::path::Path) -> ProposalDispatcher {
-        let writer: Arc<dyn FocusWriter> = Arc::new(MarkdownFocusWriter::new(focuses_root));
-        ProposalDispatcher::from_writer(
-            writer,
+        let store: Arc<dyn FocusStore> = Arc::new(MarkdownFocusStore::new(focuses_root));
+        ProposalDispatcher::from_store(
+            store,
             Arc::new(|| "2026-04-30T12:00:00Z".to_string()),
             Arc::new(|| "id-fixed".to_string()),
         )
