+++ crates/http-api/src/serve.rs
@@ -3,7 +3,7 @@ use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use adhd_ranch_storage::{DecisionLog, FocusRepository, FocusWriter, ProposalQueue};
+use adhd_ranch_storage::{DecisionLog, FocusStore, ProposalQueue};
 use tokio::net::TcpListener;
 use tokio::sync::oneshot;
 use tokio::task::JoinHandle;
@@ -73,8 +73,7 @@ impl Drop for ServerHandle {
 }
 
 pub async fn serve(
-    repo: Arc<dyn FocusRepository>,
-    writer: Arc<dyn FocusWriter>,
+    store: Arc<dyn FocusStore>,
     queue: Arc<dyn ProposalQueue>,
     decisions: Arc<dyn DecisionLog>,
     dispatcher: Arc<ProposalDispatcher>,
@@ -90,7 +89,7 @@ pub async fn serve(
         std::fs::write(path, addr.port().to_string())?;
     }
 
-    let app = router(repo, writer, queue, decisions, dispatcher);
+    let app = router(store, queue, decisions, dispatcher);
     let (shutdown_tx, shutdown_rx) = oneshot::channel::<()>();
     let join = tokio::spawn(async move {
         let _ = axum::serve(listener, app)
@@ -111,10 +110,7 @@ pub async fn serve(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use adhd_ranch_storage::{
-        FocusWriter, JsonlDecisionLog, JsonlProposalQueue, MarkdownFocusRepository,
-        MarkdownFocusWriter,
-    };
+    use adhd_ranch_storage::{JsonlDecisionLog, JsonlProposalQueue, MarkdownFocusStore};
     use std::fs;
     use tempfile::TempDir;
 
@@ -137,27 +133,19 @@ mod tests {
             "a",
             &focus_md("a", "Alpha", "x", "2026-04-30T12:00:00Z"),
         );
-        let repo = Arc::new(MarkdownFocusRepository::new(&focuses_root));
-        let writer: Arc<dyn FocusWriter> = Arc::new(MarkdownFocusWriter::new(&focuses_root));
+        let store: Arc<dyn FocusStore> = Arc::new(MarkdownFocusStore::new(&focuses_root));
         let queue = Arc::new(JsonlProposalQueue::new(dir.path().join("proposals.jsonl")));
         let decisions = Arc::new(JsonlDecisionLog::new(dir.path().join("decisions.jsonl")));
-        let dispatcher = Arc::new(ProposalDispatcher::from_writer(
-            writer.clone(),
+        let dispatcher = Arc::new(ProposalDispatcher::from_store(
+            store.clone(),
             Arc::new(|| "2026-04-30T12:00:00Z".to_string()),
             Arc::new(|| "id".to_string()),
         ));
         let port_file = dir.path().join("run/port");
 
-        let handle = serve(
-            repo,
-            writer,
-            queue,
-            decisions,
-            dispatcher,
-            Some(port_file.clone()),
-        )
-        .await
-        .unwrap();
+        let handle = serve(store, queue, decisions, dispatcher, Some(port_file.clone()))
+            .await
+            .unwrap();
         let port = std::fs::read_to_string(&port_file)
             .unwrap()
             .trim()
