+++ crates/storage/src/lib.rs
@@ -1,15 +1,13 @@
 pub mod atomic;
 pub mod decisions;
+pub mod focus_store;
 pub mod jsonl;
 pub mod proposals;
-pub mod repository;
 pub mod watcher;
-pub mod writer;
 
 pub use atomic::atomic_write;
 pub use decisions::{DecisionLog, DecisionLogError, JsonlDecisionLog};
+pub use focus_store::{FocusStore, FocusStoreError, MarkdownFocusStore};
 pub use jsonl::{JsonlError, JsonlLog};
 pub use proposals::{JsonlProposalQueue, ProposalQueue, QueueError};
-pub use repository::{FocusRepository, MarkdownFocusRepository, RepositoryError};
 pub use watcher::{watch_path, FocusWatcher, WatcherError};
-pub use writer::{FocusWriter, MarkdownFocusWriter, WriterError};
